development/prasanna/pytest-data-validation-509-equifax_data_validation/src/requirement.txt
pytest==7.1.3
pytest-forked==1.4.0
pytest-html==1.22.0
pytest-metadata==1.8.0
pytest-xdist==2.5.0
python-dateutil==2.8.1
pytz==2022.1
pandas==1.5.2
memory_profiler
Pympler==1.0.1
hurry==1.1
hurry.filesize==0.9
matplotlib
psycopg2-binary==2.9.3
future==0.18.2
voluptuous==0.13.1
flatten-json==0.1.13
requests==2.28.1
PyGithub==1.58.2
